nearly the final cuts of main activity
