+++ app/src/main/java/com/example/my_parking_app/CustomAdapter.kt
@@ -6,49 +6,21 @@ import android.widget.TextView
 import androidx.appcompat.view.menu.ActionMenuItemView
 import androidx.recyclerview.widget.RecyclerView
 
-class CustomAdapter(private val mList: List<ParkingList>, private val onClick: (ParkingList) -> Unit) :
-    RecyclerView.Adapter<CustomAdapter.ParkingViewHolder>(){
-
-    class ParkingViewHolder(itemView: View, val onClick: (ParkingList) -> Unit) :
-        RecyclerView.ViewHolder(itemView) {
-        private val id: TextView = itemView.findViewById(R.id.lot_id)
-        private val date: TextView = itemView.findViewById(R.id.date)
-        private val hour: TextView = itemView.findViewById(R.id.hour)
-        private var currentParking: ParkingList? = null
-
-        init {
-            itemView.setOnClickListener{
-                currentParking?.let {
-                    onClick(it)
-                }
-            }
-
-        }
-
-        fun bind(parking: ParkingList){
-            currentParking = parking
-            id.text = parking.id
-            if(parking.vacancy) {
-                date.text = parking.date
-                hour.text = parking.hour
-            }
-            else{
-                date.setText(R.string.free2)
-                hour.text = ""
-            }
-        }
-
-            }
+class CustomAdapter(private val mList: List<ParkingList>, private val onItemClicked: OnItemClicked) :
+    RecyclerView.Adapter<ParkingViewHolder>(){
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ParkingViewHolder {
         val view = LayoutInflater.from(parent.context)
             .inflate(R.layout.parking_text, parent, false)
-        return ParkingViewHolder(view, onClick)
+        return ParkingViewHolder(view)
     }
 
     override fun onBindViewHolder(holder: ParkingViewHolder, position: Int) {
         val parking = mList[position]
         holder.bind(parking)
+        holder.itemView.setOnClickListener{
+            onItemClicked.onLotClick(position)
+        }
     }
 
     override fun getItemCount(): Int {
