+++ app/src/main/java/com/example/my_parking_app/Lot_Details.kt
@@ -1,0 +1,11 @@
+package com.example.my_parking_app
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class Lot_Details : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_lot_details)
+    }
+}
