+++ app/src/main/java/com/example/my_parking_app/OnItemClicked.kt
@@ -1,0 +1,7 @@
+package com.example.my_parking_app
+
+interface OnItemClicked{
+        fun onLotClick(position: Int){
+
+        }
+}
