+++ app/src/main/java/com/example/my_parking_app/ParkingViewHolder.kt
@@ -1,0 +1,26 @@
+package com.example.my_parking_app
+
+import android.view.View
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+class ParkingViewHolder(itemView: View) :
+    RecyclerView.ViewHolder(itemView) {
+    private val id: TextView = itemView.findViewById(R.id.lot_id)
+    private val date: TextView = itemView.findViewById(R.id.date)
+    private val hour: TextView = itemView.findViewById(R.id.hour)
+
+
+
+    fun bind(parking: ParkingList){
+        id.text = parking.id
+        if(parking.vacancy) {
+            date.text = parking.date
+            hour.text = parking.hour
+        }
+        else{
+            date.setText(R.string.free2)
+            hour.text = ""
+        }
+    }
+
+    }
